CRUD 인증기능 추가

+++ src/main/java/com/example/newsfeed/service/MenuService.java
@@ -4,7 +4,9 @@ import com.example.newsfeed.dto.CommonResponseDto;
 import com.example.newsfeed.dto.MenuRequestDto;
 import com.example.newsfeed.dto.MenuResponseDto;
 import com.example.newsfeed.entity.Menu;
+import com.example.newsfeed.entity.User;
 import com.example.newsfeed.repository.MenuRepository;
+import com.example.newsfeed.userdetails.UserDetailsImpl;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Sort;
@@ -12,10 +14,8 @@ import org.springframework.stereotype.Service;
 
 import java.lang.reflect.Array;
 import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-import java.util.Optional;
+import java.util.*;
+import java.util.concurrent.RejectedExecutionException;
 
 @Service
 @RequiredArgsConstructor
@@ -23,8 +23,11 @@ public class MenuService {
     private final MenuRepository menuRepository;
     MenuResponseDto menuResponseDto;
     private Menu menu;
-    public MenuResponseDto post(MenuRequestDto menuRequestDto) {
+    private User user;
+    public MenuResponseDto post(MenuRequestDto menuRequestDto, UserDetailsImpl userDetails) {
+        user = userDetails.getUser();
         menu = new Menu(menuRequestDto);
+        menu.setUser(user);
         // 회원 인증 기능 추가
         menu = menuRepository.save(menu);
         return new MenuResponseDto(menu);
@@ -40,20 +43,46 @@ public class MenuService {
             return menuResponseDtoList;
     }
 
+    public MenuResponseDto getMenuByMenyId(Long menuId){
+        menu = menuRepository.findById(menuId)
+                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 게시글 ID 입니다."));
+        return new MenuResponseDto(menu);
+    }
+
+    public List<MenuResponseDto> getMenuByUserId(Long userId) {
+        List<Menu> menuList = menuRepository.findMenuByUserId(userId);
+        List<MenuResponseDto> menuResponseDtoList = new ArrayList<>();
+        for (Menu m : menuList) {
+            menuResponseDto = new MenuResponseDto(m);
+            menuResponseDtoList.add(menuResponseDto);
+        }
+        return menuResponseDtoList;
+    }
+
     @Transactional
-    public MenuResponseDto updateMenu(MenuRequestDto menuRequestDto, Long menuId) {
+    public MenuResponseDto updateMenu(MenuRequestDto menuRequestDto, Long menuId, UserDetailsImpl userDetails) {
+        user = userDetails.getUser();
         menu = getMenu(menuId);
-        menu.updateMenu(menuRequestDto);
-        menu.setModifiedAt(LocalDateTime.now());
+        if(user.getUsername().equals(menu.getUser().getUsername())) {
+            menu.updateMenu(menuRequestDto);
+            menu.setModifiedAt(LocalDateTime.now());
+        }else{
+            throw new RejectedExecutionException("작성자만 수정할 수 있습니다.");
+        }
+
         return new MenuResponseDto(menu);
     }
 
     @Transactional
-    public void deleteMenu(Long menuId) {
+    public void deleteMenu(Long menuId, UserDetailsImpl userDetails) {
+        user = userDetails.getUser();
         menu = getMenu(menuId);
-        menuRepository.delete(menu);
+        if(user.getUsername().equals(menu.getUser().getUsername())){
+            menuRepository.delete(menu);
+        }else{
+            throw new RejectedExecutionException("작성자만 삭제할 수 있습니다.");
+        }
     }
-
 
     public Menu getMenu(Long menuId) {
         return menuRepository.findById(menuId)
